+++ kartoteka-ui-gtk/src/ui/mod.rs
@@ -1,1 +1,2 @@
 pub mod app_window;
+pub mod styles;